add component notices
add every notice component in general component instead page
