+++ src/components/Notices/Notices.styled.js
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: row;
+  margin-bottom: 20px;
+`;
